test(server): split support-feed render-contract tests for repo-shape cap
The repo-shape 600-line source cap (scripts/check_max_lines.py, CI "Repo
shape checks"; evidence: CI run 29365666345 on the reviewed revision
e2e2df553) rejected tests/integration/test_support_feed.py at 650 lines.

Relocate the deploy-time task-render contract section (MERGE_JQ/ASSERT_JQ
extraction helpers and the 7 render tests) byte-identically into the
sibling tests/integration/test_support_feed_deploy_render.py. No test
logic changes, renames, or reordering; only the module header/imports of
each file are adjusted. Total collected tests remain 19 (12 + 7), all
green. Bounded amendment noted in the promoted B1 contract.
